requirements.txt
aiohttp==3.7.4
astroid==2.5.1
async-timeout==3.0.1
attrs==20.3.0
build==0.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
cogwatch==2.1.0
colorama==0.4.4
discord==1.0.1
discord.py==1.6.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
multidict==5.1.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pep517==0.9.1
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.7.2
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
watcher==0.2.1
watchgod==0.7
wrapt==1.12.1
yarl==1.6.3
